@@ -1,5 +1,5 @@
 // Code generated by protoc-gen-go-hashpb. DO NOT EDIT.
-// protoc-gen-go-hashpb (devel)
+// protoc-gen-go-hashpb
 // Source: internal/pb/all_types.proto
 
 package pb
